Move ERC20 dynamic CAL tests into their own folder
