+++ libs/ledgerjs/packages/hw-app-eth/tests/ERC20/ERC20-CAL-OK.unit.test.ts
@@ -4,8 +4,8 @@ import {
   openTransportReplayer,
   RecordStore,
 } from "@ledgerhq/hw-transport-mocker";
-import Eth from "../src/Eth";
-import ledgerService from "../src/services/ledger";
+import Eth from "../../src/Eth";
+import ledgerService from "../../src/services/ledger";
 
 const txHex =
   "f86a0e8506a2bb7d008301512c94005d1123878fc55fbd56b54c73963b234a64af3c80b844a9059cbb00000000000000000000000082ec3523f8a722694ca217ebfd95efbcadad77ee000000000000000000000000000000000000000000000002b5e3af16b1880000018080";
@@ -14,7 +14,7 @@ describe("ERC20 dynamic cal", () => {
   describe("ERC20 is in local CAL", () => {
     test("should be successfully signin transaction from dynamic CAL", async () => {
       const apdusBuffer = await fs.readFile(
-        path.resolve("./tests/erc20-transactions/ERC20-OK.apdus"),
+        path.resolve("./tests/fixtures/apdus/ERC20-OK.apdus"),
         "utf-8"
       );
       const transport = await openTransportReplayer(
@@ -42,7 +42,7 @@ describe("ERC20 dynamic cal", () => {
 
     test("should be successfully signin transaction from local CAL", async () => {
       const apdusBuffer = await fs.readFile(
-        path.resolve("./tests/erc20-transactions/ERC20-OK.apdus"),
+        path.resolve("./tests/fixtures/apdus/ERC20-OK.apdus"),
         "utf-8"
       );
       const transport = await openTransportReplayer(
