+++ libs/ledgerjs/packages/hw-app-eth/tests/fixtures/apdus/ERC20-KO.apdus
